feat: Introduce video trimming
This commit introduces the ability to trim videos, which allows users to specify a start and end time for the video that will be downloaded.
The format for specifying the trim range is as follow:

start-end

Where start and end are floats indicating the start and end timestamps in seconds (start and end are treated as floats).

Further, the concept of a DownloadQuery has been introduced to pass around information about a download more conveniently. DownloadQuery contains the Uri and an OptionsSet object.

The Downloaders have been refactored to accommodate these changes and to improve their structure and logic. VideoDownloaderBase classes now return DownloadResult objects instead of a tuple which includes the OutPath of the file and the date of the video.

The DownloadValidator has also been modified to include a method for validating the trim option, which checks if the times specified are within the valid range and exits the program with an error message if the times are out of range.

Finally, a corresponding change has been made in the DownloadFactory class to accommodate the new DownloadOptions. The code now uses a DownloadQuery object to pass around the Uri and OptionSet more conveniently.