feat(RootCommand): add version check.
Previously, if you ran `dis -v`, you would get the following error: "Error: Object reference not set to an instance of an object."

The issue lies in the fact that `Spectre.Console.Cli` helper, by default, will always include ``-h|--help` and `-v|--version` regardless of the situation. Currently, there isn't a robust logic to handle the version scenario effectively (especially in my case), so I had to implement a workaround to address it.