treewide: fetch medata data once, `GetDuration` -> `FetchMetadata`
Previously, we would fetch metadata once the TUI starts and then again
after starting a download. Obviously, this is unwanted behavior and a
waste of time

Also I replace the `GetDuration` with a more rich and generic
`FetchMetadata`, which gives us all the metadata. If in the future we
need something more than just duration, this allows us to avoid having
to create more methods